refactor(backend) : app.js / server.js 분리
