+++ backend/src/app.js
@@ -1,0 +1,12 @@
+const express = require("express");
+const cors = require("cors");
+
+const app = express();
+app.use(cors());
+app.use(express.json());
+
+app.get("/api/health", (req, res) => {
+    res.send("✅Server is running!");
+});
+
+module.exports = app;
