+++ backend/src/server.js
@@ -1,18 +1,10 @@
-const express = require("express");
-const mongoose = require("mongoose");
-const cors = require("cors");
 require("dotenv").config();
-
-const app = express();
-app.use(cors());
-app.use(express.json());
-
-app.get("/api/health", (req, res) => {
-    res.send("✅Server is running!");
-});
+const mongoose = require("mongoose");
+const app = require('./app');
 
 const PORT = process.env.PORT || 5000;
 const MONGO_URI = process.env.MONGO_URI;
+
 mongoose.connect(MONGO_URI)
     .then(() => {
         console.log("✅ MONGODB Connected.");
